fix: compile warnings and use generated cmake file (#1)

+++ src/include/hazy/hazy.h
@@ -36,7 +36,7 @@ void hazyInit(Hazy* self, size_t capacity, struct ImprintAllocator* allocator,
               struct ImprintAllocatorWithFree* allocatorWithFree, HazyConfig config, Clog log);
 void hazyReset(Hazy* self);
 void hazyUpdate(Hazy* self);
-int hazyUpdateAndCommunicate(Hazy* self, struct DatagramTransport* socket);
+ssize_t hazyUpdateAndCommunicate(Hazy* self, struct DatagramTransport* socket);
 int hazyRead(Hazy* self, uint8_t* data, size_t capacity);
 int hazyWrite(Hazy* self, const uint8_t* data, size_t octetCount);
 void hazySetConfig(Hazy* self, HazyConfig config);
